아이템, 스코어 시스템 구현 완료
아이템이 사라진 후 5초에 한번 랜덤 위치 아이템생성,
스코어, 하이스코어 시스템 적용
임시 UI 생성

+++ Assets/Scripts/Obstacles/Pumping.cs
@@ -41,7 +41,7 @@ public class Pumping : Opstacle
         // 새로운 코루틴 시작
         currentBounceCoroutine = StartCoroutine(PumpingCoroutine());
 
-        GameManager.Instance.AddScore(GameManager.Instance.Scoreweight);
+        GameManager.Instance.AddScore((int)GameManager.Instance.Scoreweight);
     }
 
     private IEnumerator PumpingCoroutine()
